Use activemq/mirror as source to remote_file.
The old source syntax is using pattern matching expecting all remote
apache mirrors to use this format. Not all mirrors are using this exact
mirror path format and many mirrors will break.  We fix this by
requiring the user to select a mirror and an exact download URL to a
.tar.gz instead.